pyproject.toml
[project]
name = "shorts-automation"
version = "0.1.0"
description = "🎬 AI-powered YouTube Shorts automation with LangGraph + MCP"
authors = [{ name = "Your Name", email = "your@email.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "boto3>=1.34.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "moviepy>=1.0.3",
    "elevenlabs>=1.0.0",
    "openai-whisper>=20231117",
    "yt-dlp>=2024.1.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "praw>=7.7.0",
    "aiohttp>=3.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
